Fix building ices under x86_64.
--HG--
branch : dbserver2

+++ ices/src/in_vorbis.c
@@ -44,7 +44,7 @@ typedef struct {
 } ices_vorbis_in_t;
 
 /* -- static prototypes -- */
-static int ices_vorbis_readpcm (input_stream_t* self, size_t len,
+static ssize_t ices_vorbis_readpcm (input_stream_t* self, size_t len,
 				int16_t* left, int16_t* right);
 static int ices_vorbis_close (input_stream_t* self);
 static void in_vorbis_parse (input_stream_t* self);
@@ -133,7 +133,7 @@ ices_vorbis_open (input_stream_t* self, char* buf, size_t len)
   return 0;
 }
 
-static int
+static ssize_t
 ices_vorbis_readpcm (input_stream_t* self, size_t olen, int16_t* left,
 		     int16_t* right)
 {
